requirements.txt
# Core dependencies
typer[all]==0.9.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Documentation
rich==13.5.2  # For better CLI output (used by typer)